src/graph/requirements.txt
fastapi 
uvicorn
langgraph 
langchain-core 
typing-extensions 
httpx 
python-dotenv 
langchain-ollama 
ipython